requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
openai>=1.40.0
python-dotenv>=1.0.1
httpx
